apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "onlook-api"
version = "0.0.0"
description = "FastAPI backend scaffold for OnlookAI and Attune Compliance"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14,<2.0",
  "fastapi>=0.115,<1.0",
  "passlib[bcrypt]>=1.7,<2.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic-settings>=2.7,<3.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "python-multipart>=0.0.20,<1.0",
  "redis>=5.2,<6.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.32,<1.0"
]

[project.optional-dependencies]
dev = [
  "mypy>=1.14,<2.0",
  "pytest>=8.3,<9.0",
  "ruff>=0.8,<1.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

